# Use double VectorTableLookup on ARM in ProbabilisticMap