--- a/src/libraries/System.Private.CoreLib/src/System/IndexOfAnyValues/ProbabilisticMap.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/IndexOfAnyValues/ProbabilisticMap.cs
@@ -172,10 +172,19 @@ namespace System.Buffers
             Vector128<byte> bitPositions = Vector128.ShuffleUnsafe(Vector128.Create(0x8040201008040201).AsByte(), highNibble);
 
             Vector128<byte> index = values & Vector128.Create((byte)VectorizedIndexMask);
-            Vector128<byte> bitMaskLower = Vector128.ShuffleUnsafe(charMapLower, index);
-            Vector128<byte> bitMaskUpper = Vector128.ShuffleUnsafe(charMapUpper, index - Vector128.Create((byte)16));
-            Vector128<byte> mask = Vector128.GreaterThan(index, Vector128.Create((byte)15));
-            Vector128<byte> bitMask = Vector128.ConditionalSelect(mask, bitMaskUpper, bitMaskLower);
+            Vector128<byte> bitMask;
+
+            if (AdvSimd.Arm64.IsSupported)
+            {
+                bitMask = AdvSimd.Arm64.VectorTableLookup((charMapLower, charMapUpper), index);
+            }
+            else
+            {
+                Vector128<byte> bitMaskLower = Vector128.ShuffleUnsafe(charMapLower, index);
+                Vector128<byte> bitMaskUpper = Vector128.ShuffleUnsafe(charMapUpper, index - Vector128.Create((byte)16));
+                Vector128<byte> mask = Vector128.GreaterThan(index, Vector128.Create((byte)15));
+                bitMask = Vector128.ConditionalSelect(mask, bitMaskUpper, bitMaskLower);
+            }
 
             return ~Vector128.Equals(bitMask & bitPositions, Vector128<byte>.Zero);
         }