Contributes to #84328

When I vectorized this path in #80963, I ran benchmarks on ARM with the initial version of the PR, but later regressed it with https://github.com/dotnet/runtime/pull/80963/commits/22953700f9e0976e4a00e34ccf69a19cf346e1a1 by using the `>>>` operator, which apparently isn't turned into `AdvSimd.ShiftRightLogical`. The first commit 3bfd7d3f82c8256c2d7d7ce9371de679d961df96 fixes that, changing perf from ~36 µs back to ~3.7 µs on [this benchmark](https://github.com/dotnet/runtime/pull/80963#issue-1551653409). cc: @tannergooding

The change to use `VectorTableLookup` with two table vectors further improves performance:

Before
|         Method |     Mean |     Error |
|--------------- |---------:|----------:|
| IndexOfAnyProb | 3.776 µs | 0.0024 µs |

After
|         Method |     Mean |     Error |
|--------------- |---------:|----------:|
| IndexOfAnyProb | 2.576 µs | 0.0011 µs |