@danmoseley 2023-04-22
> the >>> operator, which apparently isn't turned into AdvSimd.ShiftRightLogical. 

Bug?