@a74nh 2023-04-24
Code looks good to me. Out of curiosity, do you know roughly how much of the performance gain comes from the `VectorTableLookup` and how much comes from the `ShiftRightLogical` ?