@tannergooding 2023-04-24
> but later regressed it with https://github.com/dotnet/runtime/commit/22953700f9e0976e4a00e34ccf69a19cf346e1a1 by using the >>> operator, which apparently isn't turned into AdvSimd.ShiftRightLogical

Can you log a bug for this. Looks like a simple fix as we should just need to recognize the operators in addition to the named methods: https://github.com/dotnet/runtime/blob/main/src/coreclr/jit/hwintrinsicarm64.cpp#L1520-L1555