@MihaZupan 2023-04-24
> how much of the performance gain comes from the VectorTableLookup

The 3.776 µs => 2.576 µs change on this benchmark is entirely due to `VectorTableLookup`.

> Can you log a bug for this.

#85257