@kunalspathak 2023-04-24
> The 3.776 µs => 2.576 µs change on this benchmark is entirely due to `VectorTableLookup`.

That's 30% gain. Good to see that.